Emails: Move feature lists component to a better place
